refactor(providers): remove unused native usage policy
