+++ extensions/moonshot/index.ts
@@ -2,7 +2,6 @@ import { createProviderApiKeyAuthMethod } from "openclaw/plugin-sdk/provider-aut
 // Moonshot plugin entrypoint registers its OpenClaw integration.
 import { defineSingleProviderPluginEntry } from "openclaw/plugin-sdk/provider-entry";
 import { buildOpenAICompatibleReplayPolicy } from "openclaw/plugin-sdk/provider-model-shared";
-import { applyMoonshotNativeStreamingUsageCompat } from "./api.js";
 import { moonshotMediaUnderstandingProvider } from "./media-understanding-provider.js";
 import { wrapMoonshotStream } from "./native-video.js";
 import { applyMoonshotConfig, applyMoonshotConfigCn } from "./onboard.js";
@@ -47,8 +46,6 @@ export default defineSingleProviderPluginEntry({
       allowExplicitBaseUrl: true,
       liveModelDiscovery: true,
     },
-    applyNativeStreamingUsageCompat: ({ providerConfig }) =>
-      applyMoonshotNativeStreamingUsageCompat(providerConfig),
     normalizeResolvedModel: (ctx) =>
       ({
         ...ctx.model,
