+++ src/agents/models-config.applies-config-env-vars.test.ts
@@ -34,7 +34,6 @@ vi.mock("./provider-auth-aliases.js", () => ({
 // These planner tests exercise no plugin-owned auth policy. Keep their exact
 // provider markers local instead of loading the bundled plugin/runtime catalog.
 vi.mock("../plugins/provider-runtime.js", () => ({
-  applyProviderNativeStreamingUsageCompatWithPlugin: () => undefined,
   normalizeProviderConfigWithPlugin: providerRuntimeMocks.normalizeProviderConfigWithPlugin,
   resolveProviderConfigApiKeyWithPlugin: providerRuntimeMocks.resolveProviderConfigApiKeyWithPlugin,
   resolveExternalAuthProfilesWithPlugins: () => [],
