+++ src/agents/models-config.plan.ts
@@ -15,7 +15,6 @@ import {
   type ExistingProviderConfig,
 } from "./models-config.merge.js";
 import {
-  applyNativeStreamingUsageCompat,
   enforceSourceManagedProviderSecrets,
   normalizeProviderCatalogModelsForConfig,
   normalizeProviders,
@@ -303,9 +302,7 @@ async function planOpenClawModelsJsonWithDeps(
       sourceSecretDefaults: context.sourceConfigForSecrets.secrets?.defaults,
       secretRefManagedProviders,
     }) ?? normalizedMergedProviders;
-  const finalProviders = applyNativeStreamingUsageCompat(
-    filterWritableProviders(secretEnforcedProviders),
-  );
+  const finalProviders = filterWritableProviders(secretEnforcedProviders);
   const splitProviders = splitProvidersByPluginOwner({
     providers: finalProviders,
     pluginMetadataSnapshot: context.pluginMetadataSnapshot,
