+++ src/agents/models-config.providers.ts
@@ -9,5 +9,4 @@ export {
   normalizeProviders,
 } from "./models-config.providers.normalize.js";
 export type { ProviderConfig } from "./models-config.providers.secrets.js";
-export { applyNativeStreamingUsageCompat } from "./models-config.providers.policy.js";
 export { enforceSourceManagedProviderSecrets } from "./models-config.providers.source-managed.js";
